pyproject.toml
[project]
name = "resume-interviewer-agent-livekit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "cerebras-cloud-sdk>=1.56.1",
    "livekit-agents[openai,silero,turn-detector]~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "pdfplumber>=0.11.7",
    "playwright>=1.55.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.1",
]
